+++ middlewares/index.js
@@ -2,5 +2,3 @@ exports.login = (req, res, next) => {
     req.login = 'web7722'
     next()
 }
-
-// {} => { login: () => {}}
