+++ views/board/index.html
@@ -1,17 +1,20 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
 </head>
+
 <body>
     <div class="wrap">
         <div class="main">
             <h1>게시판 만들기</h1>
-            <a href="/list">board</a>
+            <a href="/board/list">board</a>
         </div>
     </div>
 </body>
+
 </html>
