+++ views/board/list.html
@@ -1,14 +1,16 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
 </head>
+
 <body>
     <div id="wrap">
-        <table border = 1>
+        <table border=1>
             <tr>
                 <th>번호</th>
                 <th>제목</th>
@@ -26,7 +28,8 @@
             </tr>
             {% endfor %}
         </table>
-        <a href="/write"> 글 작성하기 </a>
+        <a href="/board/write"> 글 작성하기 </a>
     </div>
 </body>
+
 </html>
