+++ views/board/view.html
@@ -1,11 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
 </head>
+
 <body>
     <h2>View Page</h2>
 
@@ -20,8 +22,9 @@
         -->
     </div>
 
-    <a href="/list">목록가기</a>
-    <a href="/modify?index={{item.index}}">수정하기</a>
-    <a href="/delete?index={{item.index}}">삭제하기</a>
+    <a href="/board/list">목록가기</a>
+    <a href="/board/modify?index={{item.index}}">수정하기</a>
+    <a href="/board/delete?index={{item.index}}">삭제하기</a>
 </body>
+
 </html>
