+++ views/board/write.html
@@ -1,14 +1,16 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
 </head>
+
 <body>
     <h2>글 쓰기</h2>
-    <form method='post' action="/write">
+    <form method='post' action="/board/write">
         <div>
             <div> 제목 : <input type="text" name="subject"> </div>
             <div> 작성자 : <input type="text" name="name"> </div>
@@ -17,4 +19,5 @@
         <button type="submit">글 작성완료</button>
     </form>
 </body>
+
 </html>
